+++ tests/unit/web/twig/ExtensionTest.php
@@ -12,6 +12,7 @@ use Craft;
 use craft\elements\Address;
 use craft\elements\Entry;
 use craft\elements\User;
+use craft\enums\CmsEdition;
 use craft\fields\MissingField;
 use craft\fields\PlainText;
 use craft\test\TestCase;
@@ -93,11 +94,11 @@ class ExtensionTest extends TestCase
      */
     public function testCraftSystemGlobals(): void
     {
-        Craft::$app->setEdition(Craft::Pro);
-        Craft::$app->getView()->setTemplateMode(View::TEMPLATE_MODE_CP);
+        Craft::$app->edition = CmsEdition::Pro;
         $this->testRenderResult(
-            '' . Craft::$app->getEdition() . ' | ' . Craft::Solo . ' | ' . Craft::Pro,
-            Craft::$app->getEdition() . ' | 0 | 1'
+            implode(',', [CmsEdition::Solo->value, CmsEdition::Pro->value]),
+            '{{ [CraftSolo, CraftPro]|join(",") }}',
+            templateMode: View::TEMPLATE_MODE_CP,
         );
     }
 
@@ -138,8 +139,6 @@ class ExtensionTest extends TestCase
      */
     public function testElementGlobals(): void
     {
-        Craft::$app->getView()->setTemplateMode(View::TEMPLATE_MODE_SITE);
-
         $this->testRenderResult(
             'A global set | A different global set',
             '{{ aGlobalSet }} | {{ aDifferentGlobalSet }}'
@@ -1187,12 +1186,17 @@ class ExtensionTest extends TestCase
      * @param string $expectedString
      * @param string $renderString
      * @param array $variables
+     * @param string $templateMode
      * @throws LoaderError
      * @throws SyntaxError
      */
-    protected function testRenderResult(string $expectedString, string $renderString, array $variables = [])
-    {
-        $result = $this->view->renderString($renderString, $variables);
+    protected function testRenderResult(
+        string $expectedString,
+        string $renderString,
+        array $variables = [],
+        string $templateMode = View::TEMPLATE_MODE_SITE,
+    ) {
+        $result = $this->view->renderString($renderString, $variables, $templateMode);
         self::assertSame(
             $expectedString,
             $result